refactor(agents): move Translator implementation out of assemble
The translator is an Agent with no assemble/DOM dependencies; keep the
real implementation under agents and drop the assemble-owned module plus
the empty re-export wrapper.
